RABSW-742 Control lifecycle of PV and PVC in lustre-fs-operator (#4)
Control the lifecycle of the PVC and PV by placing a finalizer on the
LustreFileSystem resource.  When deleting the LustreFileSystem resource,
explicitly delete the PV.  Let garbage collection delete the PVC when the
LustreFileSystem resource is deleted.

Define printer columns for the LustreFileSystem resource, to allow more
friendly 'kubectl get' output.

Add tests.
Do not orphan a PV in the LustreFileSystem finalizer.