RABSW-751 lustre-fs-operator should allow mgsnid to specify a hostname (#6)
Adjust the webhook to allow a hostname to be specified in the MgsNID.

Add an imagePullSecret patch for the daemonset.  Remove the 'manifests'
dependency from the 'deploy' target in the Makefile, so a customer doesn't
have to make that dependency work.

Add a sample LustreFileSystem resource and a sample Pod definition that
references the PVC created via that resource.