Generate the conversions for the v1alpha1 spoke
Update v1alpha1/groupversion_info.go to add a variable that is required
by the generated conversion routines.

Add a v1alpha1/doc.go to tell the conversion-gen tool to generate conversion
routines for that API, and to tell it that v1beta1 is the hub.

Add conversion-gen tool, and generate the conversions.  Add the conversion
library and its fuzz test.

Add the v1alpha1/conversion* spoke conversion library and self-test.

Signed-off-by: Dean Roehrich <dean.roehrich@hpe.com>