Add watch for namespaces (#50)
The Lustrefilesystem resource can get into odd situations when
namespaces are deleted. This leads to stale data in the Spec/Status
fields since nothing changes in the spec and no operator reconcile
occurs. Once this happens, PVCs will not be recreated once the namespace
comes back. This can happen when undeploying and redeploying nnf-dm when
a lustrefilesystem resource is active.

To remedy this, any change in namespaces (e.g. when nnf-dm is deployed
and the namespace is created) will trigger the reconciler and any
missing PV/PVCS will then be created.

Tested by:
1. Created a lustrefilesystem resource
2. nnf-dm then automatically adds itself to the list of namespaces in
   that resource
3. Undeploy nnf-dm (and delete namespace) - this removes the PVC
4. Redeploy nnf-dm - the PVC will be recreated mounted

Signed-off-by: Blake Devcich <blake.devcich@hpe.com>